rename: SettingConfirmAlert -> UserInformation, 마이 페이지 서버 컴포넌트로 변경
