+++ src/__tests__/components/my/UserInformation.test.tsx
@@ -1,13 +1,17 @@
 import {render, screen, fireEvent} from "@testing-library/react"
-import Component from "@/components/my/SettingConfirmAlert"
+import Component from "@/components/my/UserInformation"
 import Store from "@/store/setting-auth"
 import type {TProps as TPortalProps} from "@/components/Portal"
 import type {TProps as TSecretInputProps} from "@/components/SecretInput"
 
-const SettingConfirmAlert = ({setChecked = jest.fn()}: {setChecked?: () => void}) => {
+const UserInformation = ({
+  setChecked = jest.fn(),
+}: {
+  setChecked?: () => void
+}) => {
   return (
     <Store.Provider value={{checked: false, setChecked}}>
-      <Component isAlerting onClose={jest.fn()} />
+      <Component />
     </Store.Provider>
   )
 }
@@ -29,13 +33,32 @@ jest.mock("../../../components/SecretInput.tsx", () => ({
     <input className="password" {...props} />
   ),
 }))
+jest.mock("../../../components/Loading.tsx", () => ({
+  __esModule: true,
+  default: () => <div>loading...</div>,
+}))
 
-describe("SettingConfirmAlert", () => {
-  it("내용과 버튼을 올바르게 렌더링한다.", async () => {
+describe("UserInformation", () => {
+  it("설정 버튼을 올바르게 렌더링한다.", () => {
+    // given, when
+    render(<UserInformation />)
+
+    const settingButton = screen.getByRole("button", {name: /설정/})
+
+    // then
+    expect(settingButton).toBeInTheDocument()
+  })
+
+  it("설정 버튼을 클릭하면 내 설정 접근을 확인하는 Alert을 렌더링한다.", async () => {
     // given
-    render(<SettingConfirmAlert />)
+    render(<UserInformation />)
+
+    const settingButton = screen.getByRole("button", {name: /설정/})
 
     // when
+    fireEvent.click(settingButton)
+
+    // then
     const title = await screen.findByText(/내 설정 접근/)
     const content = await screen.findByText(
       /보안을 위해 현재 비밀번호를 입력해 주세요/
@@ -43,7 +66,6 @@ describe("SettingConfirmAlert", () => {
     const cancelButton = await screen.findByRole("button", {name: "취소"})
     const confirmButton = await screen.findByRole("button", {name: "확인하기"})
 
-    // then
     expect(title).toBeInTheDocument()
     expect(content).toBeInTheDocument()
     expect(cancelButton).toBeInTheDocument()
@@ -55,7 +77,10 @@ describe("SettingConfirmAlert", () => {
     // given
     const setCheckedMock = jest.fn()
 
-    render(<SettingConfirmAlert setChecked={setCheckedMock} />)
+    render(<UserInformation setChecked={setCheckedMock} />)
+
+    const settingButton = screen.getByRole("button", {name: /설정/})
+    fireEvent.click(settingButton)
 
     const confirmButton = await screen.findByRole("button", {name: "확인하기"})
