+++ src/app/my/page.tsx
@@ -1,21 +1,12 @@
-"use client"
-
-import {useState} from "react"
-import {NeedLoggedIn, Header, UserInformation} from "@/components"
+import {NeedLoggedIn, Header} from "@/components"
 import {
+  UserInformation,
   RollingPaperParticipation,
   ProjectsInProgressing,
   ReceivedRollingPapers,
-  SettingConfirmAlert,
 } from "@/components/my"
 
 const MyPage = () => {
-  const [isAlerting, setIsAlerting] = useState(false)
-
-  function handleIsAlerting() {
-    setIsAlerting(!isAlerting)
-  }
-
   return (
     <>
       <NeedLoggedIn />
@@ -24,22 +15,12 @@ const MyPage = () => {
         <Header />
 
         <main className="main f-center">
-          <UserInformation
-            right={
-              <button
-                className="profile-edit xxxs radius-sm mt-2"
-                onClick={handleIsAlerting}
-              >
-                ⚙️ 설정
-              </button>
-            }
-          />
+          <UserInformation />
           <RollingPaperParticipation />
           <ProjectsInProgressing />
           <ReceivedRollingPapers />
         </main>
       </div>
-      <SettingConfirmAlert isAlerting={isAlerting} onClose={handleIsAlerting} />
     </>
   )
 }
