+++ src/components/my/index.ts
@@ -3,7 +3,7 @@ import AllReceivedRollingPapers from "./AllReceivedRollingPapers"
 import ProjectsInProgressing from "./ProjectsInProgressing"
 import ReceivedRollingPapers from "./ReceivedRollingPapers"
 import RollingPaperParticipation from "./RollingPaperParticipation"
-import SettingConfirmAlert from "./SettingConfirmAlert"
+import UserInformation from "./UserInformation"
 
 export {
   AllProjects,
@@ -11,5 +11,5 @@ export {
   ProjectsInProgressing,
   ReceivedRollingPapers,
   RollingPaperParticipation,
-  SettingConfirmAlert,
+  UserInformation,
 }
